% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{get_factor_num}
\alias{get_factor_num}
\title{Determines how many factors are suitable}
\usage{
get_factor_num(df, value_cols = NULL, figs_path = NULL)
}
\description{
Determines how many factors are suitable
}
